1. Fixed the bug 1521 -- Speedline: Moving point doesn't put point at cursor
We use itk::ExtractImageFilter in ActionSpeedline.cc to extract the slice from 3D image. This filter is implemented by multithread. Hence, the order of ActionSpeedline finishing can not be guaranteed in order (in moving points). To solve this, the atomiccounter is used to drop the previous mouse events when the later one comes in. Now it has been much improved. Sometimes the moving point still cannot completely be under cursor like polyline tool. My guess is Qt thread drop some points when the system is too busy due to fastmarching costing a lot of CPU (The CPU usage reaches 70%). 

git-svn-id: https://gforge.sci.utah.edu/svn/seg3d2/trunk@1615 305cdda6-5ce1-45b3-a98d-dfc68c8b3305