Fixed inconsistency between dock window visibility and state variables.
Before this fix, if the user set a dock window to be visible by right
clicking on a dock titlebar and selecting it, the state engine was not
made aware of this and the window menu would still show the dock
window as being hidden.

Now, whenever a QtCustomDockWidget or QtCustomDialog is shown, through
any means, the state engine will be updated accordingly.



git-svn-id: https://gforge.sci.utah.edu/svn/seg3d2/trunk@1742 305cdda6-5ce1-45b3-a98d-dfc68c8b3305