Changing seg3d to use Qt dialogs instead of native.
This gets rid of segfaults in Ubuntu precise.  See the original bug report:
https://gforge.sci.utah.edu/gf/project/seg3d2/tracker/?action=TrackerItemEdit&tracker_id=710&tracker_item_id=772

jeoren checked in the following fix for this:
https://gforge.sci.utah.edu/svn/seg3d2/trunk@1708

However, this stil causes segfaults.  The actual problem most likely lies in Seg3D's linking
in of a version of libfreetype that conflicts with Ubuntu's gtk-2.  See the following link for
more information.

http://www.qtcentre.org/archive/index.php/t-37767.html