Fixed bug in isosurface generation.
If the number of faces is over 1,000,000 on a Z-iteration, it is
broken into multiple parts, but if there are no more faces after the
split, then another, empty part, is added, which later causes a crash
(at least on windows) when uploading the vertex buffer.  This fix
prevents the creation of the empty part.



git-svn-id: https://gforge.sci.utah.edu/svn/seg3d2/trunk@2043 305cdda6-5ce1-45b3-a98d-dfc68c8b3305